+++ Routing.php
@@ -1,7 +1,5 @@
 <?php
 
-require_once 'src/controllers/DefaultController.php';
-
 class Routing {
     public static $routes;
