+++ public/views/register.php
@@ -25,7 +25,7 @@
     </head>
     <body class="flex-row-center-center">
         <div class="auth-container">
-            <a href="/dashboard" class="logo">
+            <a href="/" class="logo">
                 <span class="material-symbols-outlined logo__image">
                     movie
                 </span>
